requirements_pythonanywhere.txt
# Requirements pour PythonAnywhere
# Base dependencies
python-slugify==8.0.4
Pillow==11.0.0
argon2-cffi==23.1.0
whitenoise==6.8.2

# Django
django==5.0.10
django-environ==0.11.2
django-model-utils==5.0.0
django-allauth[mfa]==65.3.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-jazzmin==3.0.1
django-cors-headers==4.6.0

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0

# MySQL connector (pour PythonAnywhere)
mysqlclient==2.2.0

# Production
gunicorn==23.0.0
